+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## Unreleased
 
 ### Improvements
+- After critical status of service health checks next check is run after 1s,2s,3s... until 10s or first pass of health checks.
+- Enabled autoreload of Armada API in development Vagrant
+
+# Changelog
+
+## Unreleased
+
+### Improvements
 - Enabled autoreload of Armada API in development Vagrant